requirements.txt
fastapi==0.109.0
uvicorn==0.27.0.post1
python-jose==3.3.0
passlib==1.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1